django_backend/requirements.txt
# Core Dependencies
Django==5.0
djangorestframework==3.14
django-cors-headers==4.3.1
mongoengine==0.27.0
pymongo==4.6.1

# ML Dependencies
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
xgboost==2.0.0

# API Documentation
drf-spectacular==0.27.0

# Development & Testing
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0

# Production
gunicorn==21.2.0